refactor : 트랜잭션이 필요한 로직들을 Service 계층으로 이동
