+++ backend/banking/src/main/java/com/togg/banking/member/application/MemberService.java
@@ -1,0 +1,23 @@
+package com.togg.banking.member.application;
+
+import com.togg.banking.member.domain.Member;
+import com.togg.banking.member.domain.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Service
+public class MemberService {
+
+    private final MemberRepository memberRepository;
+
+    public Member findByEmail(String email) {
+        return memberRepository.getByEmail(email);
+    }
+
+    public Member findByRefreshToken(String refreshToken) {
+        return memberRepository.getByRefreshToken(refreshToken);
+    }
+}
